Finished all exercises in part 4 of the course.

+++ 4/blogilista/controllers/blogs.js
@@ -4,18 +4,9 @@ const User = require('../models/user')
 const jwt = require('jsonwebtoken')
 
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer')){
-    return authorization.substring(7)
-  }
-  return null
-}
-
-
 blogsRouter.get('/', async (request, response) => {
 
-  const blogs = await Blog.find({}).populate('user', {username: 1, name: 1, id: 1})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs.map(blog => blog.toJSON()))
 
 })
@@ -23,13 +14,12 @@ blogsRouter.get('/', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
 
   const body = request.body
-  const token = getTokenFrom(request)
-  
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id){
-    return response.status(401).json({ error: 'token missing or invalid'})
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
   }
-  
+
   const user = await User.findById(decodedToken.id)
 
   const blog = new Blog({
@@ -48,12 +38,28 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid.' })
+  }
+  const blogToRemove = await Blog.findById(request.params.id)
+
+  if(!blogToRemove){
+    return response.status(404).json({ error: 'blog does not exist' })
+  }
+
+  if (blogToRemove.user.toString() !== decodedToken.id.toString()){
+    return response.status(401).json({ error: 'only user who created the blog is allowed to delete it.'})
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
+
 blogsRouter.put('/:id', async (request, response) => {
-  console.log("Put request:")
+
   const body = request.body
 
   const updatedBlog = {
